+++ src/utils/types.ts
@@ -1,0 +1,19 @@
+export interface RowData {
+  [key: string]: string | undefined;
+  columnA: string;
+  columnB: string;
+  columnC: string;
+  columnD: string;
+  columnE?: string;
+  columnF?: string;
+  columnG?: string;
+  columnH?: string;
+  columnI?: string;
+  columnJ?: string;
+  columnK?: string;
+  columnL?: string;
+  columnM?: string;
+  columnN?: string;
+  columnO?: string;
+  columnP?: string;
+}
